backup/ReceiptAnalysis_20260203/backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
Pillow>=10.3,<11
requests>=2.31,<3
pydantic>=2.6,<3
python-dotenv>=1.0,<2
djangorestframework-simplejwt>=5.3,<6
